requirements-replit.txt
flask==2.3.3
flask-cors==4.0.0
torch==2.0.1+cpu
torchvision==0.15.2+cpu
pillow==10.0.0
--extra-index-url https://download.pytorch.org/whl/cpu